actually node-resolver server not needed at all
