--- services/node/config.mjs
@@ -1,1 +1,0 @@
-export const NODE_RESOLVER_PORT = 8082;
